+++ Changelog.md
@@ -3,4 +3,5 @@
 ## v0.6.0 (unreleased)
 
 - Add a `quic.Config` option for QUIC versions
+- Add a `quic.Config` option to request truncation of the connection ID from a server
 - Various bugfixes